Move files to services.models submodule
